+++ lluvia/cpp/core/include/lluvia/core/node/ComputeNodeDescriptor.h
@@ -9,8 +9,8 @@
 #define LLUVIA_CORE_NODE_COMPUTE_NODE_DESCRIPTOR_H_
 
 #include "lluvia/core/node/Node.h"
+#include "lluvia/core/node/PushConstants.h"
 #include "lluvia/core/Parameter.h"
-#include "lluvia/core/PushConstants.h"
 #include "lluvia/core/types.h"
 
 #include <array>
