+++ lluvia/cpp/core/src/node/ComputeNode.cpp
@@ -9,14 +9,14 @@
 
 #include "lluvia/core/buffer/Buffer.h"
 #include "lluvia/core/CommandBuffer.h"
-#include "lluvia/core/node/ComputeNodeDescriptor.h"
 #include "lluvia/core/error.h"
 #include "lluvia/core/image/Image.h"
 #include "lluvia/core/image/ImageView.h"
 #include "lluvia/core/Interpreter.h"
+#include "lluvia/core/node/ComputeNodeDescriptor.h"
+#include "lluvia/core/node/PushConstants.h"
 #include "lluvia/core/Object.h"
 #include "lluvia/core/Program.h"
-#include "lluvia/core/PushConstants.h"
 
 #include "lluvia/core/vulkan/Device.h"
